refactor: encapsulate logic to set Db name and pass
it allows obtaining the DatabaseName quickly from other app places
